Add files via upload
Despite offering many options for detailed analysis of the scatter fields, the default version of "regionSizeDistribution" does not provide the distribution of entrained air volume along the wake. Understanding the bubble generation and contribution in the mixed-phase region requires analysis of this parameter. Hence, the calculation of entrained air volume was added to the utility.